Basic Menu Functionality
I put in the switch state handlers, added two new classes for ui
buttons, and made the basic transitions from the main menu to the class
select, to the game available.  Need to figure out how to increase the
size of the window, add 'back' buttons, and clean up code for
instanciating buttons.