General code clean up, and actually making a list of user selected units
The private list<unit> variable called userSelectedUnits, actually holds
all the units that the user selected now!!! I think that we should make
a class called UnitManager or something along those lines, so that not
everything is in a jumbled mess like is now. We need something we can
call from the Game1 class and be able to do everything from there. This
will also help with networking in the future. I will start to work after
I submit our code so far