Game Manager class
The list of units that the player has picked from the menu is now a
property of the GameManager class, the menuHandler has gone inside of
the GameManager class, and a whole lot of other good stuff. There is a
basic draw method in the Unit class now. This is good stuff people, but
still a work in progress.