Made it so that a keyboard state now gets passed into the update method
This will help us out in the future when we need to check for abilities
and stuff. I also removed IDamageable because we never use it and
everyone said that it was ok. Created an ability class, but i don't
really know what to do with it right now. Added in Rate of Fire in the
unit class and to the text file