Some major changes to the enemies and stuff
I changed the quit button to ESC, and pause to P, because that was
annoying me, you can change it back if you want.  I made an enemy class,
a graph class, and a node class. I am going to be using these for
pathfinding and other AI features like shooting whenever a unit is
inside the range. Which still dont work Hansoo