Remove Bad Collision handling code, and implement control point
I set up the control point so now you can capture it, and it will change
color to represent who controls it.  In order to do this, I had to
change some of the code that was really screwing up our collisions
detecting.  I also changed the spawn locations so that no unit spawns in
a wall now.  I'm still curious why the units disappear upon hitting a
wall.