Implements Spawn Locations, and Respawn functionality
Changed of code in the unit class to handle when the unit is dead and
alive.  Set up the respawn countdown, and the spawn location for each
unit is set upon the game start.  Still need to implement the correct
health value upon respawn, and the ability to switch classes.